[update] 아트그램 상세모달 리펠토링 및 read 부분 관련구성
