+++ src/features/artgram/ArtgramForm.jsx
@@ -11,20 +11,20 @@ import { useGetimgurl } from "../../hooks/artgram/useGetimgurl";
 
 
 function ArtgramForm() {
-  // 비동기 통신을 위하 커스텀 훅(리액트 쿼리)
+  // 비동기 통신을 위하 커스텀 훅(리액트 쿼리)  ------------------------------------------------------------------ //
   const [postArtgrams] = usePostartgram();
 
   // Form의 input state 관리
   const [formState, setFormState, handleInputChange] = useFormInput();
 
-  // Drag&Drop files state 관리 및 화면에 미리보기 제어
+  // Drag&Drop files state 관리 및 화면에 미리보기 제어   ----------------------------------------------------- //
   const [files, setFiles, getRootProps, getInputProps] = useDropzoneinput()
   useEffect(() => {
     // 마운트 해제시, 데이터 url 취소
     return () => files.forEach((file) => URL.revokeObjectURL(file.preview));
   }, []);
 
-  // Drag&Drop state(files)를 AWS S3에 업로드하여 url 받아내고, newImageUrls state에 입력하기 
+  // Drag&Drop state(files)를 AWS S3에 업로드하여 url 받아내고, newImageUrls state에 입력하기   ---------------- //
   const [s3imgurlhandle] = useGetimgurl(files)
 
   const handleSubmit = (event) => {
@@ -35,10 +35,11 @@ function ArtgramForm() {
     setFiles([])
     setFormState({})
   };
-
+  // -------------------------------------------------------------------------------------------------- //
   return (
     <>
       <Flex as="form" onSubmit={handleSubmit} fd="column" gap="10">
+        {/* input:Atom을 활용한 molecule --------------------------------------------------------------- */} 
         {createArtgramInputList.map((input, index) => (
           <Input
             key={index}
@@ -52,7 +53,7 @@ function ArtgramForm() {
           />
         ))}
 
-        {/* <section className="container"> */}
+        {/* dropzone 라이브러리를 활용한 type:file 업로드 -------------------------------------------------- */}
         <ArtgramFormparts.Section {...getRootProps({ className: "dropzone" })}>
           <input {...getInputProps()} />
           <ArtgramFormparts.DragIcon>
@@ -62,6 +63,7 @@ function ArtgramForm() {
             Drag & drop some files here, or click to select files
           </ArtgramFormparts.DragText>
         </ArtgramFormparts.Section>
+        {/* dropzone 라이브러리를 활용, 업로드된 이미지에 대한 미리보기 ------------------------------------------- */}
         <ArtgramFormparts.ThumbsContainer>
           {files &&
             files.map((file) => (
@@ -77,6 +79,7 @@ function ArtgramForm() {
               </ArtgramFormparts.Thumb>
             ))}
         </ArtgramFormparts.ThumbsContainer>
+        {/* form 태그의  submit 버든 -------------------------------------------------------------------- */}
         <input type="submit" value="등록하기" />
       </Flex>
     </>
